+++ src/melclouderv.js
@@ -18,6 +18,7 @@ class MelCloudErv extends EventEmitter {
         const restFulEnabled = config.restFulEnabled;
         const mqttEnabled = config.mqttEnabled;
         const deviceInfoFile = `${prefDir}/${accountName}_Device_${deviceId}`;
+        this.refreshInterval = config.refreshInterval;
 
         this.axiosInstanceGet = axios.create({
             method: 'GET',
@@ -324,7 +325,7 @@ class MelCloudErv extends EventEmitter {
                 await new Promise(resolve => setTimeout(resolve, 1000));
                 this.emit('checkDeviceState');
             } catch (error) {
-                this.emit('error', `check info, ${error}, check again in 60s.`);
+                this.emit('error', `check info, ${error}, check again in ${this.refreshInterval / 1000}s.`);
                 this.checkDeviceInfo();
             };
         }).on('checkDeviceState', async () => {
@@ -418,7 +419,7 @@ class MelCloudErv extends EventEmitter {
 
                 this.checkDeviceInfo();
             } catch (error) {
-                this.emit('error', `check device state error, ${error}, check again in 60s.`);
+                this.emit('error', `check device state error, ${error}, check again in ${this.refreshInterval / 1000}s.`);
                 this.checkDeviceInfo();
             };
         });
@@ -427,7 +428,7 @@ class MelCloudErv extends EventEmitter {
     };
 
     async checkDeviceInfo() {
-        await new Promise(resolve => setTimeout(resolve, 65000));
+        await new Promise(resolve => setTimeout(resolve, this.refreshInterval));
         this.emit('checkDeviceInfo');
     };